논리 오류 수정, 회원 삭제 전 비밀번호 검증 로직 추가

+++ src/main/java/com/ngplpresentation/ngpl_backend/service/MemberService.java
@@ -41,7 +41,7 @@ public class MemberService {
         if (pwd.length() < PWD_MIN_LENGTH || pwd.length() > PWD_MAX_LENGTH)
             throw new GeneralException(Status.BAD_REQUEST_ARGUMENT_PWD_MIN_MAX);
 
-        if (pwd.equals(confpwd))
+        if (!pwd.equals(confpwd))
             throw new GeneralException(Status.BAD_REQUEST_ARGUMENT_PWD_NOT_SAME);
 
         if (memberRepository.findByUserId(userId).isPresent())
@@ -77,6 +77,9 @@ public class MemberService {
     public void deleteMember(DeleteRequest req) {
         Member member = memberRepository.findByUserId(req.getUserId())
                 .orElseThrow(() -> new GeneralException(Status.BAD_REQUEST_NON_EXISTENCE_MEMBER));
+
+        if (!member.getPassword().equals(req.getPassword()))
+            throw new GeneralException(Status.BAD_REQUEST_ARGUMENT_PWD_NOT_SAME);
 
         memberRepository.deleteById(member.getId());
     }
